Finish chapter 10
Making Docking Stations Release Bikes

∆ Add test for #DockingStation.release_bike. This tests to check a
DockingStation releases a new Bike each time the method is called. Test
also expects that Bike to be #working?
∆ Refactor #DockingStation.release_bike to generate a new instance of
Bike class each time it is called (with #Bike.new). This is the simplest
way to ensure the test passes for now.
∆ Refactor #Bike.working? to now have a hardcoded return of 'true' each
time it is called. This is also the simplest solution for passing the
test.