Complete Red-Green stages of chapter 12
Raising Exceptions

∆ Add test that, upon calling #release_bike, expects an error to be
raised if that state of the DockingStation is such that @bike instance
variable returns nil. This is assuming that if the @bike instance
variable is nil, it would mean the DockingStation is empty.

∆ Add if logic to the #release_bike method in DockingStation so the
previously described test now passes. This prevents the DockingStation
from releasing new Bike objects when it is empty. Currently concerned
about the multiple responsibilities of #release_bike.

∆ Refactor 'can release bike' test in 'docking_station_spec'. This is
because the test was expecting new Bike objects to be created upon a
call of #release_bike. The recent changes to the code described in the
preceeding points have meant that this now raises an error, so the
test needed to make sure @bike is changed from nil, then #release_bike
returns the bike that is saved in the state of the DockingStation object
it has been called upon.