Complete Red-Green stages of chapter 13
Limiting Capacity

∆ Add test 'raises an error if capacity reached'. This ensures bikes
cannot be docked if the DockingStation already contains as many bikes
as it can hold.

∆ Add a call to #capacity_error from #dock(bike), both in
DockingStation. This is to leave the raising of an error to a seperate
method and ensure #dock(bike) simply checks and sets state. Still feel
it may be possible to seperate out the checking and the setting, this
feels like two resposibilities.

∆ Add two private methods to DockingStation class, #empty? and #full?.
This starts to extrct the checking of state from #release_bike and