Pass 'customer_spec' test 'can view menu items'
∆ Add #avaliable_dishes and #takeaway to customer class. This inits a
customer with a takeaway that they will be ordering and provides a
means for them to check which dishes are avaliable.

∆ Update sequence diagram so the Customer asks the Takeaway for the
returns a #dishes_list and a #prices_list. The Takeaway then handles
formatting these two objects to return a #formatted_menu to the
Customer.

Add further functionality not needed to pass above stated test

∆ Create the actual Takeaway class and add an anemic method
unit tests.